Retrieve Reasons for Not Achieving Good Status

Retrieves details of Reasons for Not Achieving Good (RNAG) status and Reasons For Failure (RFF) from EA Catchment Data Explorer site. Data can be retrieved by specifying waterbody id (WBID), Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD). Start year (startyr) and end year (endyr) allow specific timeranges to be downloaded. For Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD) level downloads, waterbody type can also be specified to allow extraction of specific waterbody types (River, Lake etc). Data are presented at the level of individual elements that are the reasons for not achieving good status.

get_rnag(ea_name = NULL, column = NULL, startyr = NULL,
  endyr = NULL, type = NULL)

Arguments

ea_name
A string representing the description (name for OC, MC or RBD level downloads or WBID for individual waterbodies) of the features to be extracted. For example to extract data for the whole of the Humber RBD, this would be "Humber"; also see examples. Must be an exact match to the values used in the EA database. Use the search_names function to search for specific values.

column
The column to be searched. Possible options are WBID (waterbody id), OC (Operational Catchment), MC (Management Catchment) and RBD (River Basin District)

startyr
The data can be extracted for specific years using the startyr and endyr arguments. If only startyr is specified this extracts for a particular year. If no years are specified all years are returned. RNAG data are only available from 2013 onwards.

endyr
The data can be extracted for specific years using the startyr and endyr arguments. The endyr should only be specified if startyr is also included, otherwise it returns an error. RNAG data are only available from 2013 onwards.

type
Type of waterbody to be extracted. For Operational/Management catchment level or RBD level queries, the data can also be subset by waterbody type. Possible values are River, Lake, GroundWaterBody, TransitionalWater or CoastalWater.

Value

An object of class cde_df containing the details of the Reasons for Not Achieving Good Status for the specified combination of criteria. For details of the meaning of the the different columns returned, see https://docs.ropensci.org/cde/articles/cde-output-reference.html.

Examples

# get all RNAG issues identified for waterbody GB112071065700
get_rnag("GB112071065700", "WBID")

#>  river_basin_district management_catchment operational_catchment
#>            North West               Ribble            Big Ribble
#>            North West               Ribble            Big Ribble
#> With an additional 27 columns of data. 
#> Row values may be truncated to fit console. 

# get the RNAG issues for Lakes in the Humber RBD, between
# 2013 and 2014
get_rnag(ea_name="Humber", column="RBD", startyr=2013, endyr=2014, type="Lake")

#> Processing data from zipfile...

#>  river_basin_district management_catchment operational_catchment water_body
#>                Humber Trent Valley Staffor Penk Rivers and Lakes Belvide Re
#>                Humber Trent Valley Staffor Penk Rivers and Lakes Belvide Re
#>                Humber Trent Valley Staffor Penk Rivers and Lakes Belvide Re
#>                Humber Trent Valley Staffor Penk Rivers and Lakes Belvide Re
#>                Humber Trent Valley Staffor Penk Rivers and Lakes Belvide Re
#>                Humber Trent Valley Staffor Blithe Rivers and Lak Blithfield
#>                Humber Trent Valley Staffor Blithe Rivers and Lak Blithfield
#>                Humber      Aire and Calder          Calder Upper Widdop Res
#>                Humber      Aire and Calder          Calder Upper Widdop Res
#>                Humber      Aire and Calder          Calder Upper Walshaw De
#> With an additional 139 rows and 22 columns of data. 
#> Row values may be truncated to fit console. 

# get the RNAG issues for Rivers in the Avon Warwickshire
# Management Catchment in 2015
get_rnag(ea_name="Avon Warwickshire", column="MC", startyr=2015, type="River")

#> Processing data from zipfile...

#>  river_basin_district management_catchment operational_catchment water_body
#>                Severn    Avon Warwickshire  Avon - Midlands West Harvington
#>                Severn    Avon Warwickshire  Avon - Midlands West Harvington
#>                Severn    Avon Warwickshire  Avon - Midlands West Stoulton B
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#>                Severn    Avon Warwickshire  Avon - Midlands West Tirle Broo
#> With an additional 103 rows and 22 columns of data. 
#> Row values may be truncated to fit console.