Retrieve Objectives

Retrieves details of objectives set for waterbodies in terms of predicted classification from EA Catchment Data Explorer site. Data can be retrieved by specifying waterbody id (WBID), Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD). Start year (startyr) and end year (endyr) allow specific timeranges to be downloaded. For Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD) level downloads, waterbody type can also be specified to allow extraction of specific waterbody types (River, Lake etc).

get_objectives(ea_name = NULL, column = NULL,
  level = "Overall Water Body", year = NULL, type = NULL)

Arguments

ea_name
A string representing the description (name) of the features to be extracted. For example to extract data for the whole of the Humber RBD, this would be "Humber"; also see examples. Must be an exact match to the values used in the EA database. Use the search_names function to search for specific values.

column
The column to be searched. Possible options are WBID (waterbody id), OC (Operational Catchment), MC (Management Catchment) and RBD (River Basin District)

level
The level within the WFD quality classification that objectives have been set at. Defaults to 'Overall Water Body'. See information in the Vignette for possible values.

year
The year that objectives are set for, either 2015, 2021 or 2027. If not given then objectives for all years are returned. Note that objectives may not be set for all years.

type
Type of waterbody to be extracted. For Operational/Management catchment level or RBD level queries, the data can also be subset by waterbody type. Possible values are River, Lake, GroundWaterBody, TransitionalWater or CoastalWater.

Value

A data frame containing the details of the objectives set for the specified set of waterbodies.

Examples

# get all objectives set for waterbody GB112071065700

# NOT RUN {
get_objectives("GB112071065700", "WBID")
# }

# get the objectives set for Lakes in the Humber RBD, for the year 2021

# NOT RUN {
get_objectives("Humber", "RBD", year=2021, type="Lake")
# }

# get the objectives set for Rivers in the Avon Warwickshire
# Operational Catchment in relation to Chemical status

# NOT RUN {
get_objectives("Avon Warwickshire", "MC", level="Chemical", type = "River")
# }