Retrieve Reasons for Not Achieving Good Status

Retrieves details of Reasons for Not Achieving Good (RNAG) status and Reasons For Failure (RFF) from EA Catchment Data Explorer site. Data can be retrieved by specifying waterbody id (WBID), Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD). Start year (startyr) and end year (endyr) allow specific timeranges to be downloaded. For Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD) level downloads, waterbody type can also be specified to allow extraction of specific waterbody types (River, Lake etc). Data is presented at the level of individual elements that are the reasons for not achieving good status.

get_rnag(ea_name = NULL, column = NULL, startyr = NULL,
  endyr = NULL, type = NULL)

Arguments

ea_name
A string representing the description (name) of the features to be extracted. For example to extract data for the whole of the Humber RBD, this would be "Humber"; also see examples. Must be an exact match to the values used in the EA database. Use the search_names function to search for specific values.

column
The column to be searched. Possible options are WBID (waterbody id), OC (Operational Catchment), MC (Management Catchment) and RBD (River Basin District)

startyr
The data can be extracted for specific years using the startyr and endyr arguments. If only startyr is specified this extracts for a particular year. If no years are specified all years are returned. RNAG data are only available from 2013 onwards.

endyr
The data can be extracted for specific years using the startyr and endyr arguments. The endyr should only be specified if startyr is also included, otherwise it is ignored and all years are returned. RNAG data are only available from 2013 onwards.

type
Type of waterbody to be extracted. For Operational/Management catchment level or RBD level queries, the data can also be subset by waterbody type. Possible values are River, Lake, GroundWaterBody, TransitionalWater or CoastalWater.

Value

A data frame containing the details of the Reasons for Not Achieving Good Status for the specified combination of criteria.

Examples

# get all RNAG issues identified for waterbody GB112071065700

# NOT RUN {
get_rnag("GB112071065700", "WBID")
# }

# get the RNAG issues for Lakes in the Humber RBD, between
# 2013 and 2014

# NOT RUN {
get_rnag("Humber", "RBD", startyr = 2013, endyr = 2014, type = "Lake")
# }

# get the RNAG issues for Rivers in the Avon Warwickshire
# Management Catchment in 2015

# NOT RUN {
get_rnag("Avon Warwickshire", "MC", startyr = 2015, type = "River")
# }