Search database of site names

Searches the listing of EA monitoring sites to find rows that contain the string provided. Can search by name (name), Management Catchment (MC), Operational Catchment (OC) or River Basin District (RBD). The search is done on a local copy of the waterbody listing rather than connecting to the EA site.

search_names(string = NULL, column = NULL)

Arguments

string
The search string to be matched. Will match whole or partial strings in the column values.

column
The column to be searched. Possible options are name, OC (Operational Catchment), MC (Management Catchment) and RBD (River Basin District)

Value

A data frame containing the details of all the sites that match the search string (full or partial matches) in the column specified.

Examples

# search for Management Catchments containing "Kennet"
search_names("Kennet", "MC")

#>                WBID                                                       name
#> 2589 GB106039017360                           Inkpen Stream (source to Kennet)
#> 2590 GB106039017300                 West End Brook (tributary of Foudry Brook)
#> 2591 GB106039017380                        Foudry Brook (West End Brook to M4)
#> 2592 GB106039023120                                             Clayhill Brook
#> 2593 GB106039023173                  Middle Kennet (Marlborough to Hungerford)
#> 2594 GB106039017410                                           Burghfield Brook
#> 2595 GB106039017370               Shalbourne  (source to Kennet at Hungerford)
#> 2596 GB106039017280                         Enborne (Source to downstream A34)
#> 2597 GB106039017310               Enborne (downstream A34 to Burghclere Brook)
#> 2598     GB30642611                         Farnham Flint or Englefield Lagoon
#> 2599 GB106039017210                                             Penwood Stream
#> 2600 GB106039017250                      Ecchinswell Brook (source to Enborne)
#> 2601     GB30642622                                 Ameys Lake or Theale Lakes
#> 2602 GB106039023200                                                  Aldbourne
#> 2603     GB30643126                                               Milford Lake
#> 2604 GB106039017190                                           Silchester Brook
#> 2605 GB106039023140                                      Kennet and Holy Brook
#> 2606 GB106039023180                                                         Og
#> 2607 GB106039017261             Enborne (Burghclere Brook to Kingsclere Brook)
#> 2608 GB106039017390             Kennet and Avon Canal and Dun above Hungerford
#> 2609 GB106039023220                               Lambourn (Source to Newbury)
#> 2610 GB106039023171                                Upper Kennet to Marlborough
#> 2611 GB106039017200                                            Baughurst Brook
#> 2612 GB106039017220                       Kingsclere Brook (Source to Enborne)
#> 2613 GB106039023210                                               Winterbourne
#> 2614 GB106039017340                                              Lower Enborne
#> 2615 GB106039023174                      Middle Kennet (Hungerford to Newbury)
#> 2616 GB106039017420         Kennet (Lambourn confluence to Enborne confluence)
#> 2617 GB106039023141                                                 Holy Brook
#> 2618 GB106039017350                                                  Upper Dun
#> 2619 GB106039017230 Earlstone Stream and  Burghclere Brook (source to Enborne)
#> 2620 GB106039017430                                           Froxfield Stream
#> 2621     GB30643001                                          Wasing Wood Ponds
#>       type     OC                     MC    RBD
#> 2589 River Kennet Kennet and tributaries Thames
#> 2590 River Kennet Kennet and tributaries Thames
#> 2591 River Kennet Kennet and tributaries Thames
#> 2592 River Kennet Kennet and tributaries Thames
#> 2593 River Kennet Kennet and tributaries Thames
#> 2594 River Kennet Kennet and tributaries Thames
#> 2595 River Kennet Kennet and tributaries Thames
#> 2596 River Kennet Kennet and tributaries Thames
#> 2597 River Kennet Kennet and tributaries Thames
#> 2598  Lake Kennet Kennet and tributaries Thames
#> 2599 River Kennet Kennet and tributaries Thames
#> 2600 River Kennet Kennet and tributaries Thames
#> 2601  Lake Kennet Kennet and tributaries Thames
#> 2602 River Kennet Kennet and tributaries Thames
#> 2603  Lake Kennet Kennet and tributaries Thames
#> 2604 River Kennet Kennet and tributaries Thames
#> 2605 River Kennet Kennet and tributaries Thames
#> 2606 River Kennet Kennet and tributaries Thames
#> 2607 River Kennet Kennet and tributaries Thames
#> 2608 River Kennet Kennet and tributaries Thames
#> 2609 River Kennet Kennet and tributaries Thames
#> 2610 River Kennet Kennet and tributaries Thames
#> 2611 River Kennet Kennet and tributaries Thames
#> 2612 River Kennet Kennet and tributaries Thames
#> 2613 River Kennet Kennet and tributaries Thames
#> 2614 River Kennet Kennet and tributaries Thames
#> 2615 River Kennet Kennet and tributaries Thames
#> 2616 River Kennet Kennet and tributaries Thames
#> 2617 River Kennet Kennet and tributaries Thames
#> 2618 River Kennet Kennet and tributaries Thames
#> 2619 River Kennet Kennet and tributaries Thames
#> 2620 River Kennet Kennet and tributaries Thames
#> 2621  Lake Kennet Kennet and tributaries Thames

# search for River Basin Districts containing "Tweed"

# NOT RUN {
search_names("Tweed", "RBD")
# }